+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Added Thermo Fisher Qubit Flex adapter 
+- Added requirement document for Thermo Fisher Qubit Flex adapter
 
 ### Fixed
 
@@ -24,8 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added electrophoresis and spectrophotometry parser requirement templates to /docs
-- Added Thermo Fisher Scientific Qubit Flex adapter 
-- Added requirement document for Thermo Fisher Scientific Qubit Flex adapter
 
 ### Fixed